requirements.txt
amqp==2.5.2
anyjson==0.3.3
APScheduler==3.6.3
asgiref==3.2.5
bangla==0.0.2
beautifulsoup4==4.8.2
billiard==3.6.3.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
get==2019.4.13
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
numpy==1.18.2
pandas==1.0.3
plotly==4.5.4
post==2019.4.13
public==2019.4.13
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
wincertstore==0.2
zipp==3.1.0
